HW types and variables - changed int to double in HW_04 and removed empty variables in HW_07; + added zip in upload folder

+++ Prep_Course/01_Types_and_Variables/HomeWork/HW_04/Program.cs
@@ -11,11 +11,11 @@ namespace HW_04
         static void Main(string[] args)
         {
             // input
-            int first = int.Parse(Console.ReadLine());
-            int second = int.Parse(Console.ReadLine());
+            double first = double.Parse(Console.ReadLine());
+            double second = double.Parse(Console.ReadLine());
 
             // calculation
-            int result = (first + second + Math.Abs(first - second)) / 2;
+            double result = (first + second + Math.Abs(first - second)) / 2;
 
             // output
             Console.WriteLine(result);
